title仮完成、ライフサイクルシステムもまとまってきてwave管理もいい感じ、この調子

+++ Game/source/engine/core/game_timer.cpp
@@ -11,7 +11,7 @@ namespace sdl_engine
    {
       // デルタタイムを計算
       u64 tick_time = SDL_GetPerformanceCounter();
-      _delta_time   = static_cast<f64>( tick_time - _last_time ) / SDL_GetPerformanceFrequency();
+      _delta_time   = static_cast<f32>( tick_time - _last_time ) / SDL_GetPerformanceFrequency();
       _last_time    = tick_time;
 
       // 経過時間を加算
@@ -23,7 +23,7 @@ namespace sdl_engine
       // fps計算
       if ( _elapsed_time >= 1.0 )
       {
-         _fps          = _frame_count / _elapsed_time;
+         _fps          = static_cast<u32>( _frame_count / _elapsed_time );
          _frame_count  = 0;
          _elapsed_time = 0.0;
       }
